finally kill the TypeChecker 🎉
