+++ spy/vm/typechecker.py
@@ -25,306 +25,6 @@ if TYPE_CHECKING:
 #     binary operators)
 DispatchKind = Literal['single', 'multi']
 
-def maybe_blue(*colors: Color) -> Color:
-    """
-    Return 'blue' if all the given colors are blue, else 'red'
-    """
-    if set(colors) == {'blue'}:
-        return 'blue'
-    else:
-        return 'red'
-
-
-class TypeChecker:
-    vm: 'SPyVM'
-    w_func: W_ASTFunc
-    funcef: ast.FuncDef
-    expr_types: dict[ast.Expr, tuple[Color, W_Type]]
-    expr_conv: dict[ast.Expr, W_Func]
-    opimpl: dict[ast.Node, W_Func]
-    locals_types_w: dict[str, W_Type]
-
-
-    def __init__(self, vm: 'SPyVM', w_func: W_ASTFunc) -> None:
-        self.vm = vm
-        self.w_func = w_func
-        self.funcdef = w_func.funcdef
-        self.expr_types = {}
-        # XXX: expr_conv nowadays is used only for typechecking locals. Maybe
-        # we should use a better system?
-        self.expr_conv = {}
-        self.opimpl = {}
-        self.locals_types_w = {}
-        self.declare_arguments()
-
-    def typecheck_local(self, expr: ast.Expr, name: str) -> None:
-        assert name in self.locals_types_w
-        got_color, w_got_type = self.check_expr(expr)
-
-    def name2sym_maybe(self, expr: ast.Expr) -> Optional[Symbol]:
-        """
-        If expr is an ast.Name, return the corresponding Symbol.
-        Else, return None.
-        """
-        if isinstance(expr, ast.Name):
-            return self.funcdef.symtable.lookup_maybe(expr.id)
-        return None
-
-    def check_stmt(self, stmt: ast.Stmt) -> None:
-        magic_dispatch(self, 'check_stmt', stmt)
-
-    def check_expr(self, expr: ast.Expr) -> tuple[Color, W_Type]:
-        """
-        Compute the STATIC type of the given expression
-        """
-        if expr in self.expr_types:
-            return self.expr_types[expr]
-        else:
-            color, w_type = magic_dispatch(self, 'check_expr', expr)
-            self.expr_types[expr] = color, w_type
-            return color, w_type
-
-    def check_many_exprs(self,
-                         prefixes: list[str],
-                         exprs: Sequence[ast.Expr | str]
-                         ) -> tuple[list[Color], list[W_OpArg]]:
-        assert len(prefixes) == len(exprs)
-        colors = []
-        args_wop = []
-        last_loc = None
-        for i, (prefix, expr) in enumerate(zip(prefixes, exprs)):
-            if isinstance(expr, str):
-                # HACK HACK HACK: we need a loc but we don't have any. We just
-                # use the last_loc. This works only as far as this doesn't
-                # happen at the first iteration, which is good enough in
-                # practice.
-                #
-                # Ultimately this happens because astr.GetAttr.attr is of type
-                # 'str'. Probably we should just turn it into a "real" Expr,
-                # so that it will be automatically and transparetly converted
-                # into a W_OpArg
-                assert last_loc is not None
-                loc = last_loc
-                color = 'blue'
-                wop = W_OpArg('blue', B.w_str, loc, w_val = self.vm.wrap(expr))
-            else:
-                color, w_type = self.check_expr(expr)
-                wop = W_OpArg('red', w_type, expr.loc,
-                              sym=self.name2sym_maybe(expr))
-                last_loc = expr.loc
-            colors.append(color)
-            args_wop.append(wop)
-        return colors, args_wop
-
-    # ==== statements ====
-
-    def check_stmt_Return(self, ret: ast.Return) -> None:
-        self.check_expr(ret.value)
-
-    def check_stmt_Pass(self, stmt: ast.Pass) -> None:
-        pass
-
-
-    def check_stmt_StmtExpr(self, stmt: ast.StmtExpr) -> None:
-        pass
-
-    def check_stmt_If(self, if_node: ast.If) -> None:
-        color, w_got_type = self.check_expr(if_node.test)
-
-    def check_stmt_While(self, while_node: ast.While) -> None:
-        color, w_got_type = self.check_expr(while_node.test)
-
-    def check_stmt_UnpackAssign(self, unpack: ast.UnpackAssign) -> None:
-        _, w_valuetype = self.check_expr(unpack.value)
-        if w_valuetype is not B.w_tuple:
-            t = w_valuetype.fqn.human_name
-            err = SPyTypeError(f'`{t}` does not support unpacking')
-            err.add('error', f'this is `{t}`', unpack.value.loc)
-            raise err
-
-        for i, target in enumerate(unpack.targets):
-            # we need an expression which has the type of each individual item
-            # of the tuple. The easiest way is to synthetize a GetItem
-            expr = ast.GetItem(
-                loc = unpack.value.loc,
-                value = unpack.value,
-                index = ast.Constant(
-                    loc = unpack.value.loc,
-                    value = i
-                )
-            )
-            self._check_assign(target, expr)
-
-    def check_stmt_SetAttr(self, node: ast.SetAttr) -> None:
-        _, args_wop = self.check_many_exprs(
-            ['t', 'a', 'v'],
-            [node.target, node.attr.value, node.value]
-        )
-        w_opimpl = self.vm.call_OP(OP.w_SETATTR, args_wop)
-        self.opimpl[node] = w_opimpl
-
-    def check_stmt_SetItem(self, node: ast.SetItem) -> None:
-        _, args_wop = self.check_many_exprs(
-            ['t', 'i', 'v'],
-            [node.target, node.index, node.value]
-        )
-        w_opimpl = self.vm.call_OP(OP.w_SETITEM, args_wop)
-        self.opimpl[node] = w_opimpl
-
-    # ==== expressions ====
-
-    def check_expr_Name(self, name: ast.Name) -> tuple[Color, W_Type]:
-        varname = name.id
-        sym = self.funcdef.symtable.lookup_maybe(varname)
-        if sym is None:
-            msg = f"name `{name.id}` is not defined"
-            raise SPyNameError.simple(msg, "not found in this scope", name.loc)
-        elif sym.fqn:
-            # XXX this is wrong: we should keep track of the static type of
-            # FQNs. For now, we just look it up and use the dynamic type
-            w_value = self.vm.lookup_global(sym.fqn)
-            assert w_value is not None
-            return sym.color, self.vm.dynamic_type(w_value)
-        elif sym.is_local:
-            return sym.color, self.locals_types_w[name.id]
-        else:
-            # closed-over variables are always blue
-            namespace = self.w_func.closure[sym.level]
-            w_value = namespace[sym.name]
-            assert w_value is not None
-            return 'blue', self.vm.dynamic_type(w_value)
-
-    def check_expr_Constant(self, const: ast.Constant) -> tuple[Color, W_Type]:
-        T = type(const.value)
-        assert T in (int, float, bool, NoneType)
-        if T is int:
-            return 'blue', B.w_i32
-        elif T is float:
-            return 'blue', B.w_f64
-        elif T is bool:
-            return 'blue', B.w_bool
-        elif T is NoneType:
-            return 'blue', B.w_void
-        assert False
-
-    def check_expr_StrConst(self, const: ast.StrConst) -> tuple[Color, W_Type]:
-        return 'blue', B.w_str
-
-    def check_expr_FQNConst(self, const: ast.FQNConst) -> tuple[Color, W_Type]:
-        # XXX: I think that FQNConst should remember what was its static type
-        w_val = self.vm.lookup_global(const.fqn)
-        assert w_val is not None
-        w_type = self.vm.dynamic_type(w_val)
-        return 'blue', w_type
-
-    def check_expr_BinOp(self, binop: ast.BinOp) -> tuple[Color, W_Type]:
-        w_OP = OP_from_token(binop.op) # e.g., w_ADD, w_MUL, etc.
-        colors, args_wop = self.check_many_exprs(
-            ['l', 'r'],
-            [binop.left, binop.right],
-        )
-        color = maybe_blue(*colors)
-        w_opimpl = self.vm.call_OP(w_OP, args_wop)
-        self.opimpl[binop] = w_opimpl
-        return color, w_opimpl.w_functype.w_restype
-
-    check_expr_Add = check_expr_BinOp
-    check_expr_Sub = check_expr_BinOp
-    check_expr_Mul = check_expr_BinOp
-    check_expr_Div = check_expr_BinOp
-    check_expr_Eq = check_expr_BinOp
-    check_expr_NotEq = check_expr_BinOp
-    check_expr_Lt = check_expr_BinOp
-    check_expr_LtE = check_expr_BinOp
-    check_expr_Gt = check_expr_BinOp
-    check_expr_GtE = check_expr_BinOp
-
-    def check_expr_GetItem(self, expr: ast.GetItem) -> tuple[Color, W_Type]:
-        colors, args_wop = self.check_many_exprs(
-            ['v', 'i'],
-            [expr.value, expr.index],
-        )
-        color = maybe_blue(*colors)
-        w_opimpl = self.vm.call_OP(OP.w_GETITEM, args_wop)
-        self.opimpl[expr] = w_opimpl
-        return color, w_opimpl.w_functype.w_restype
-
-    def check_expr_GetAttr(self, expr: ast.GetAttr) -> tuple[Color, W_Type]:
-        colors, args_wop = self.check_many_exprs(
-            ['v', 'a'],
-            [expr.value, expr.attr.value] # hack hack hack
-        )
-        w_opimpl = self.vm.call_OP(OP.w_GETATTR, args_wop)
-        self.opimpl[expr] = w_opimpl
-        return colors[0], w_opimpl.w_functype.w_restype
-
-    def check_expr_Call(self, call: ast.Call) -> tuple[Color, W_Type]:
-        # XXX: how are we supposed to know the color of the result if we
-        # are calling a dynamic expr?
-        # E.g.:
-        #
-        # @blue
-        # def foo(): ...
-        #
-        # @blue
-        # def bar(): ...
-        #     x: dynamic = foo
-        #     x()   # color???
-        #
-        # So far we always return red, because callop._dynamic_call_opimpl
-        # returns a w_functype of red color. But we need to think whether this
-        # is the real behavior that we want.
-
-        n = len(call.args)
-        colors, args_wop = self.check_many_exprs(
-            ['f'] + ['v']*n,
-            [call.func] + call.args
-        )
-        w_opimpl = self.vm.call_OP(OP.w_CALL, args_wop)
-        self.opimpl[call] = w_opimpl
-        w_functype = w_opimpl.w_functype
-        return w_functype.color, w_functype.w_restype
-
-    def check_expr_CallMethod(self, op: ast.CallMethod) -> tuple[Color, W_Type]:
-        n = len(op.args)
-        colors, args_wop = self.check_many_exprs(
-            ['t', 'm'] + ['v']*n,
-            [op.target, op.method.value] + op.args  # type: ignore
-        )
-        w_opimpl = self.vm.call_OP(OP.w_CALL_METHOD, args_wop)
-        self.opimpl[op] = w_opimpl
-        w_functype = w_opimpl.w_functype
-        return w_functype.color, w_functype.w_restype
-
-    def check_expr_List(self, listop: ast.List) -> tuple[Color, W_Type]:
-        w_itemtype = None
-        color: Color = 'red' # XXX should be blue?
-
-        for item in listop.items:
-            c1, w_t1 = self.check_expr(item)
-            color = maybe_blue(color, c1)
-            if w_itemtype is None:
-                w_itemtype = w_t1
-            w_itemtype = self.vm.union_type(w_itemtype, w_t1)
-        #
-        # XXX we need to handle empty lists
-        assert w_itemtype is not None
-        w_listtype = self.vm.make_list_type(w_itemtype)
-        return color, w_listtype
-
-    def check_expr_Tuple(self, tupleop: ast.Tuple) -> tuple[Color, W_Type]:
-        color: Color = 'blue'
-        for item in tupleop.items:
-            c1, w_t1 = self.check_expr(item)
-            color = maybe_blue(color, c1)
-        return color, B.w_tuple
-
-
-
-
-# ===== NEW STYLE TYPECHECKING =====
-# A lot of this code is copied&pasted from TypeChecker for now.
-# The goal is to kill the TypeChecker class eventually
 
 def typecheck_opimpl(
         vm: 'SPyVM',
